Remove unused imports and functions (pyflakes)
